// FRONTEND/chatbot-admin-panel/src/services/api/schemas/preMFATokenResponseSchema.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */

export interface PreMFATokenResponseSchema {
  status?: string;
  /** El username_ad (DNI) del usuario que requiere verificación MFA. */
  username_ad: string;
  /** Token JWT temporal para ser usado en el endpoint de verificación MFA. */
  pre_mfa_token: string;
  token_type?: string;
  message?: string;
}